moved assert group into cls.Fus
