+++ dss/util/auth/authorize.py
@@ -21,12 +21,8 @@ class Authorize(metaclass=AuthRegistry):
     def security_flow(self, *args, **kwargs):
         """
         This function maps out flow for a given security config
-        Be sure to call it from the derived class to ensure the JWT has groups set
         """
-        self.assert_required_parameters(kwargs, ['groups', 'token'])
-        groups = kwargs.get('security_groups')
-        token = kwargs.get('security_token')
-        self.assert_authorized_group(groups, token)
+        raise NotImplementedError()
 
     def assert_required_parameters(self, provided_params: dict, required_params: list):
         """
@@ -42,8 +38,3 @@ class Authorize(metaclass=AuthRegistry):
                 raise DSSException(500, title, err)
         return
 
-    def assert_authorized_group(self, group: typing.List[str], token: dict) -> None:
-        if token.get(Config.get_OIDC_group_claim()) in group:
-            return
-        logger.info(f"User not in authorized group: {group}, {token}")
-        raise DSSForbiddenException()
